feat(freedesktop): Implement freedesktop notifications
Refactored all freedesktop.org code to a single file
